Move schools views into separate views sub-package
