+++ anyway/flask_app.py
@@ -31,10 +31,14 @@ from webassets.ext.jinja2 import AssetsExtension
 from werkzeug.security import check_password_hash
 from wtforms import form, fields, validators, StringField, PasswordField, Form
 
-from anyway.app import app, db
+from anyway.base import app, db
 from anyway.helpers import get_kwargs, generate_csv, generate_json, involved_data_refinement, vehicles_data_refinement, \
     get_involved_dict, get_vehicle_dict, parse_data, get_json_object, log_bad_request, post_handler, string2timestamp, \
     year2timestamp, PreferenceObject, get_dict_file
+from anyway.views.schools.api import schools_api, schools_description_api, schools_yishuvs_api, schools_names_api, \
+    injured_around_schools_api, injured_around_schools_sex_graphs_data_api, \
+    injured_around_schools_months_graphs_data_api
+from anyway.views.schools.base import schools, schools_report
 from . import utilities
 from .base import user_optional
 from .clusters_calculator import retrieve_clusters
@@ -90,6 +94,10 @@ Compress(app)
 CORS(app, resources={r"/location-subscription": {"origins": "*"}})
 
 
+@app.teardown_appcontext
+def shutdown_session(exception=None):
+    db.session.remove()
+
 
 @babel.localeselector
 def get_locale():
@@ -100,22 +108,8 @@ def get_locale():
         return lang
 
 
-@app.route('/schools', methods=["GET"])
-@user_optional
-def schools():
-    if request.method == "GET":
-        return render_template('schools_dashboard_react.html')
-    else:
-        return Response("Method Not Allowed", 405)
-
-
-@app.route('/schools-report', methods=["GET"])
-@user_optional
-def schools_report():
-    if request.method == "GET":
-        return render_template('schools_dashboard_react.html')
-    else:
-        return Response("Method Not Allowed", 405)
+app.add_url_rule('/schools', endpoint=None, view_func=schools, methods=["GET"])
+app.add_url_rule('/schools-report', endpoint=None, view_func=schools_report, methods=["GET"])
 
 
 @app.route("/markers", methods=["GET"])
@@ -179,293 +173,13 @@ def single_news_flash(news_flash_id):
     return Response(status=404)
 
 
-@app.route("/api/schools", methods=["GET"])
-@user_optional
-def schools_api():
-    logging.debug('getting schools')
-    schools = db.session.query(School).filter(not_(and_(School.latitude == 0, School.longitude == 0)), \
-                                              not_(and_(School.latitude == None, School.longitude == None))) \
-        .with_entities(School.yishuv_symbol,
-                       School.yishuv_name,
-                       School.school_name,
-                       School.longitude,
-                       School.latitude).all()
-    schools_list = [{"yishuv_symbol": x.yishuv_symbol,
-                     "yishuv_name": x.yishuv_name,
-                     "school_name": x.school_name,
-                     "longitude": x.longitude,
-                     "latitude": x.latitude} for x in schools]
-    response = Response(json.dumps(schools_list, default=str), mimetype="application/json")
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/schools-description", methods=["GET"])
-@user_optional
-def schools_description_api():
-    logging.debug('getting schools with description')
-    query_obj = db.session.query(SchoolWithDescription) \
-        .filter(not_(and_(SchoolWithDescription.latitude == 0, SchoolWithDescription.longitude == 0)), \
-                not_(and_(SchoolWithDescription.latitude == None, SchoolWithDescription.longitude == None)), \
-                or_(SchoolWithDescription.school_type == 'גן ילדים', SchoolWithDescription.school_type == 'בית ספר')) \
-        .with_entities(SchoolWithDescription.school_id,
-                       SchoolWithDescription.school_name,
-                       SchoolWithDescription.students_number,
-                       SchoolWithDescription.municipality_name,
-                       SchoolWithDescription.yishuv_name,
-                       SchoolWithDescription.geo_district,
-                       SchoolWithDescription.school_type,
-                       SchoolWithDescription.foundation_year,
-                       SchoolWithDescription.location_accuracy,
-                       SchoolWithDescription.longitude,
-                       SchoolWithDescription.latitude)
-    df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-    schools_list = df.to_dict(orient='records')
-    response = Response(json.dumps(schools_list, default=str), mimetype="application/json")
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/schools-yishuvs", methods=["GET"])
-@user_optional
-def schools_yishuvs_api():
-    logging.debug('getting schools yishuvs')
-    schools_yishuvs = db.session.query(SchoolWithDescription) \
-        .filter(not_(and_(SchoolWithDescription.latitude == 0, SchoolWithDescription.longitude == 0)), \
-                not_(and_(SchoolWithDescription.latitude == None, SchoolWithDescription.longitude == None)), \
-                or_(SchoolWithDescription.school_type == 'גן ילדים', SchoolWithDescription.school_type == 'בית ספר')) \
-        .group_by(SchoolWithDescription.yishuv_name) \
-        .with_entities(SchoolWithDescription.yishuv_name).all()
-    schools_yishuvs_list = sorted([x[0] for x in schools_yishuvs])
-    response = Response(json.dumps(schools_yishuvs_list, default=str), mimetype="application/json")
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/schools-names", methods=["GET"])
-@user_optional
-def schools_names_api():
-    logging.debug('getting schools names')
-    query_obj = db.session.query(SchoolWithDescription) \
-        .filter(not_(and_(SchoolWithDescription.latitude == 0, SchoolWithDescription.longitude == 0)), \
-                not_(and_(SchoolWithDescription.latitude == None, SchoolWithDescription.longitude == None)), \
-                or_(SchoolWithDescription.school_type == 'גן ילדים', SchoolWithDescription.school_type == 'בית ספר')) \
-        .with_entities(SchoolWithDescription.yishuv_name,
-                       SchoolWithDescription.school_name,
-                       SchoolWithDescription.longitude,
-                       SchoolWithDescription.latitude,
-                       SchoolWithDescription.school_id)
-    df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-    df = df.groupby(['yishuv_name', 'school_name', 'longitude', 'latitude']).min()
-    df = df.reset_index(drop=False)
-    schools_names_ids = df.to_dict(orient='records')
-    response = Response(json.dumps(schools_names_ids, default=str), mimetype="application/json")
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/injured-around-schools", methods=["GET"])
-@user_optional
-def injured_around_schools_api():
-    report_years = [2014, 2015, 2016, 2017, 2018]
-    logging.debug('getting injured around schools api')
-    school_id = request.values.get('school_id')
-    if school_id is not None:
-        query_obj = db.session.query(InjuredAroundSchool).filter(InjuredAroundSchool.school_id == school_id) \
-            .with_entities(InjuredAroundSchool.school_id,
-                           InjuredAroundSchool.accident_year,
-                           InjuredAroundSchool.killed_count,
-                           InjuredAroundSchool.severly_injured_count,
-                           InjuredAroundSchool.light_injured_count,
-                           InjuredAroundSchool.total_injured_killed_count,
-                           InjuredAroundSchool.rank_in_yishuv)
-        df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        if not df.empty:
-            for year in report_years:
-                if year not in list(df.accident_year.unique()):
-                    df = df.append({'school_id': df.school_id.values[0],
-                                    'accident_year': year,
-                                    'killed_count': 0,
-                                    'severly_injured_count': 0,
-                                    'light_injured_count': 0,
-                                    'total_injured_killed_count': 0,
-                                    'rank_in_yishuv': df.rank_in_yishuv.values[0]},
-                                   ignore_index=True)
-            final_list = df.to_dict(orient='records')
-            response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-            response.headers.add('Access-Control-Allow-Origin', '*')
-            return response
-        else:
-            query_obj = db.session.query(SchoolWithDescription) \
-                .filter(SchoolWithDescription.school_id == school_id) \
-                .with_entities(SchoolWithDescription.school_id)
-            df_school_id = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-            if not df_school_id.empty:
-                final_list = []
-                for year in report_years:
-                    final_list.append({'school_id': school_id,
-                                       'accident_year': year,
-                                       'killed_count': 0,
-                                       'severly_injured_count': 0,
-                                       'light_injured_count': 0,
-                                       'total_injured_killed_count': 0,
-                                       'rank_in_yishuv': None})
-                response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-                response.headers.add('Access-Control-Allow-Origin', '*')
-                return response
-
-        response = Response(status=404)
-        response.headers.add('Access-Control-Allow-Origin', '*')
-        return response
-    school_yishuv_name = request.values.get('school_yishuv_name')
-    if school_yishuv_name is not None:
-        query_obj = db.session.query(InjuredAroundSchool).filter(
-            InjuredAroundSchool.school_yishuv_name == school_yishuv_name) \
-            .with_entities(InjuredAroundSchool.school_id,
-                           InjuredAroundSchool.accident_year,
-                           InjuredAroundSchool.killed_count,
-                           InjuredAroundSchool.severly_injured_count,
-                           InjuredAroundSchool.light_injured_count,
-                           InjuredAroundSchool.total_injured_killed_count,
-                           InjuredAroundSchool.rank_in_yishuv)
-        df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        final_list = df.to_dict(orient='records')
-        if not df.empty:
-            response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-            response.headers.add('Access-Control-Allow-Origin', '*')
-            return response
-        response = Response(status=404)
-        response.headers.add('Access-Control-Allow-Origin', '*')
-        return response
-    query_obj = db.session.query(InjuredAroundSchool).with_entities(InjuredAroundSchool.school_id,
-                                                                    InjuredAroundSchool.accident_year,
-                                                                    InjuredAroundSchool.killed_count,
-                                                                    InjuredAroundSchool.severly_injured_count,
-                                                                    InjuredAroundSchool.light_injured_count,
-                                                                    InjuredAroundSchool.total_injured_killed_count,
-                                                                    InjuredAroundSchool.rank_in_yishuv)
-    df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-    final_list = df.to_dict(orient='records')
-    response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/injured-around-schools-sex-graphs-data", methods=["GET"])
-@user_optional
-def injured_around_schools_sex_graphs_data_api():
-    logging.debug('getting injured around schools sex graphs data api')
-    school_id = request.values.get('school_id')
-    if school_id is not None:
-        query_obj = db.session.query(InjuredAroundSchoolAllData) \
-            .filter(InjuredAroundSchoolAllData.school_id == school_id) \
-            .join(Sex, and_(InjuredAroundSchoolAllData.involved_sex == Sex.id,
-                            InjuredAroundSchoolAllData.markers_accident_year == Sex.year,
-                            InjuredAroundSchoolAllData.markers_provider_code == Sex.provider_code)) \
-            .with_entities(InjuredAroundSchoolAllData.school_id,
-                           Sex.sex_hebrew,
-                           func.count(InjuredAroundSchoolAllData.school_id)) \
-            .group_by(InjuredAroundSchoolAllData.school_id,
-                      Sex.sex_hebrew)
-        df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        query_obj = db.session.query(Sex) \
-            .with_entities(Sex.sex_hebrew)
-        df_sex = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        df_sex = df_sex.groupby(['sex_hebrew']).size().reset_index(name='count')
-        query_obj = db.session.query(SchoolWithDescription) \
-            .filter(SchoolWithDescription.school_id == school_id) \
-            .with_entities(SchoolWithDescription.school_id)
-        df_school_id = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        if not df.empty:
-            for sex in list(df_sex['sex_hebrew'].unique()):
-                if sex not in list(df['sex_hebrew'].unique()):
-                    df = df.append({'school_id': school_id,
-                                    'sex_hebrew': sex,
-                                    'count_1': 0},
-                                   ignore_index=True)
-            final_list = df.to_dict(orient='records')
-            response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-            response.headers.add('Access-Control-Allow-Origin', '*')
-            return response
-        else:
-            if not df_school_id.empty:
-                final_list = []
-                for sex in list(df_sex['sex_hebrew'].unique()):
-                    final_list.append({'school_id': school_id,
-                                       'sex_hebrew': sex,
-                                       'count_1': 0})
-                response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-                response.headers.add('Access-Control-Allow-Origin', '*')
-                return response
-    response = Response(status=404)
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
-
-
-@app.route("/api/injured-around-schools-months-graphs-data", methods=["GET"])
-@user_optional
-def injured_around_schools_months_graphs_data_api():
-    logging.debug('getting injured around schools months graphs data api')
-    school_id = request.values.get('school_id')
-    if school_id is not None:
-        query_obj = db.session.query(InjuredAroundSchoolAllData) \
-            .filter(InjuredAroundSchoolAllData.school_id == school_id) \
-            .join(AccidentMonth, and_(InjuredAroundSchoolAllData.markers_accident_month == AccidentMonth.id,
-                                      InjuredAroundSchoolAllData.markers_accident_year == AccidentMonth.year,
-                                      InjuredAroundSchoolAllData.markers_provider_code == AccidentMonth.provider_code)) \
-            .join(InjurySeverity, and_(InjuredAroundSchoolAllData.involved_injury_severity == InjurySeverity.id,
-                                       InjuredAroundSchoolAllData.markers_accident_year == InjurySeverity.year,
-                                       InjuredAroundSchoolAllData.markers_provider_code == InjurySeverity.provider_code)) \
-            .with_entities(InjuredAroundSchoolAllData.school_id,
-                           AccidentMonth.accident_month_hebrew,
-                           InjurySeverity.injury_severity_hebrew,
-                           func.count(InjuredAroundSchoolAllData.school_id)) \
-            .group_by(InjuredAroundSchoolAllData.school_id,
-                      AccidentMonth.accident_month_hebrew,
-                      InjurySeverity.injury_severity_hebrew)
-        df = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        query_obj = db.session.query(AccidentMonth) \
-            .with_entities(AccidentMonth.accident_month_hebrew)
-        df_month = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        df_month = df_month.groupby(['accident_month_hebrew']).size().reset_index(name='count')
-        query_obj = db.session.query(InjurySeverity) \
-            .with_entities(InjurySeverity.injury_severity_hebrew)
-        df_injury_severity = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-        df_injury_severity = df_injury_severity.groupby(['injury_severity_hebrew']).size().reset_index(name='count')
-        if not df.empty:
-            for month in list(df_month['accident_month_hebrew'].unique()):
-                for injury_severity in list(df_injury_severity['injury_severity_hebrew'].unique()):
-                    if month not in list(df.accident_month_hebrew.unique()) \
-                            or injury_severity not in list(
-                        df[df.accident_month_hebrew == month].injury_severity_hebrew.unique()):
-                        df = df.append({'school_id': school_id,
-                                        'accident_month_hebrew': month,
-                                        'injury_severity_hebrew': injury_severity,
-                                        'count_1': 0},
-                                       ignore_index=True)
-            final_list = df.to_dict(orient='records')
-            response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-            response.headers.add('Access-Control-Allow-Origin', '*')
-            return response
-        else:
-            query_obj = db.session.query(SchoolWithDescription) \
-                .filter(SchoolWithDescription.school_id == school_id) \
-                .with_entities(SchoolWithDescription.school_id)
-            df_school_id = pd.read_sql_query(query_obj.statement, query_obj.session.bind)
-            if not df_school_id.empty:
-                final_list = []
-                for month in list(df_month['accident_month_hebrew'].unique()):
-                    for injury_severity in list(df_injury_severity['injury_severity_hebrew'].unique()):
-                        final_list.append({'school_id': school_id,
-                                           'accident_month_hebrew': month,
-                                           'injury_severity_hebrew': injury_severity,
-                                           'count_1': 0})
-                response = Response(json.dumps(final_list, default=str), mimetype="application/json")
-                response.headers.add('Access-Control-Allow-Origin', '*')
-                return response
-    response = Response(status=404)
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    return response
+app.add_url_rule("/api/schools", endpoint=None, view_func=schools_api, methods=["GET"])
+app.add_url_rule("/api/schools-description", endpoint=None, view_func=schools_description_api, methods=["GET"])
+app.add_url_rule("/api/schools-yishuvs", endpoint=None, view_func=schools_yishuvs_api, methods=["GET"])
+app.add_url_rule("/api/schools-names", endpoint=None, view_func=schools_names_api, methods=["GET"])
+app.add_url_rule("/api/injured-around-schools", endpoint=None, view_func=injured_around_schools_api, methods=["GET"])
+app.add_url_rule("/api/injured-around-schools-sex-graphs-data", endpoint=None, view_func=injured_around_schools_sex_graphs_data_api, methods=["GET"])
+app.add_url_rule("/api/injured-around-schools-months-graphs-data", endpoint=None, view_func=injured_around_schools_months_graphs_data_api, methods=["GET"])
 
 
 @app.route("/charts-data", methods=["GET"])
